Link during controller instantiation
