+++ src/fastcs/controller.py
@@ -3,9 +3,11 @@ from __future__ import annotations
 from collections.abc import Iterator
 from copy import copy
 from dataclasses import dataclass
+from types import MethodType
 
-from .attributes import Attribute
+from .attributes import Attribute, AttrW, Sender
 from .cs_methods import Command, Put, Scan
+from .exceptions import FastCSException
 from .wrappers import WrappedMethod
 
 
@@ -98,12 +100,18 @@ class Controller(BaseController):
 
     def __init__(self) -> None:
         super().__init__()
+        self.link_process_tasks()
 
     async def initialise(self) -> None:
         pass
 
     async def connect(self) -> None:
         pass
+
+    def link_process_tasks(self):
+        for single_mapping in self.get_controller_mappings():
+            _link_single_controller_put_tasks(single_mapping)
+            _link_attribute_sender_class(single_mapping)
 
 
 class SubController(BaseController):
@@ -115,3 +123,39 @@ class SubController(BaseController):
 
     def __init__(self) -> None:
         super().__init__()
+
+
+def _link_single_controller_put_tasks(single_mapping: SingleMapping) -> None:
+    for name, method in single_mapping.put_methods.items():
+        name = name.removeprefix("put_")
+
+        attribute = single_mapping.attributes[name]
+        match attribute:
+            case AttrW():
+                attribute.set_process_callback(
+                    MethodType(method.fn, single_mapping.controller)
+                )
+            case _:
+                raise FastCSException(
+                    f"Mode {attribute.access_mode} does not "
+                    f"support put operations for {name}"
+                )
+
+
+def _link_attribute_sender_class(single_mapping: SingleMapping) -> None:
+    for attr_name, attribute in single_mapping.attributes.items():
+        match attribute:
+            case AttrW(sender=Sender()):
+                assert (
+                    not attribute.has_process_callback()
+                ), f"Cannot assign both put method and Sender object to {attr_name}"
+
+                callback = _create_sender_callback(attribute, single_mapping.controller)
+                attribute.set_process_callback(callback)
+
+
+def _create_sender_callback(attribute, controller):
+    async def callback(value):
+        await attribute.sender.put(controller, attribute, value)
+
+    return callback
